// src/enums/financialFormFields.enum.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-shadow */
export enum FinancialFormFieldsEnum {
  // userInfoStep
  NAME = "userInfoStep.name",
  FIRST_LAST_NAME = "userInfoStep.firstLastName",
  SECOND_LAST_NAME = "userInfoStep.secondLastName",
  PHONE = "userInfoStep.phone",
  ADDRESS = "userInfoStep.address",
  AGE = "userInfoStep.age",
  CURP = "userInfoStep1.curp",
  EMAIL = "userInfoStep1.email",
  GENDER = "userInfoStep1.gender",
  DAY = "userInfoStep1.day",
  MONTH = "userInfoStep1.month",
  YEAR = "userInfoStep1.year",

  // incomeStep
  MONTHLY_INCOME= "incomeStep.monthlyIncome",
  MONTHLY_OUTCOME= "incomeStep.monthlySpending",
  // dependantsStep
  CHILDREN_COUNT = "dependantsStep.childrenCount",
  DEPENDANTS_COUNT = "dependantsStep.dependantsCount",
  ECONOMIC_ACTIVITY = "dependantsStep1.economicActivity",
  COMPANY_NAME =  "dependantsStep1.companyName",
  EDUCATIONAL_LEVEL =  "dependantsStep1.educationalLevel",
  // userStatusStep
  CIVIL_STATUS = "userStatusStep.civilStatus",
  LIVES_WITH = "userStatusStep.livesWith",
  ASSETS = "userStatusStep1.assets",
  OTHER_ASSETS = "userStatusStep1.otherAsset",
}